backend/requirements.txt
# Core Flask and Web Framework
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-CORS>=4.0.0
Flask-JWT-Extended>=4.5.0

# Database
SQLAlchemy>=2.0.0
alembic>=1.12.0

# Security and Authentication
bcrypt>=4.0.0
cryptography>=41.0.0
PyJWT>=2.8.0

# AI and Machine Learning
openai>=1.3.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Google APIs
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# HTTP and Networking
requests==2.31.0
urllib3==2.0.7

# Data Processing
pandas==2.1.3
python-dateutil==2.8.2

# Caching and Performance
redis==5.0.1

# Environment and Configuration
python-dotenv==1.0.0

# Logging and Monitoring
structlog==23.2.0

# Development and Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0

# Production Server
gunicorn==21.2.0
gevent==23.9.1

# Utilities
click==8.1.7
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2

# Vector Database (lightweight alternatives)
chromadb==0.4.18
faiss-cpu==1.7.4

# Additional utilities for cost optimization
cachetools==5.3.2
joblib==1.3.2
